+++ src/medicines/medicines.module.ts
@@ -1,4 +1,5 @@
 import { Module } from '@nestjs/common';
+import { MedicinesCacheAdminService } from './cache/medicines-cache-admin.service';
 import { MedicinesCacheService } from './cache/medicines-cache.service';
 import { MedicinesController } from './medicines.controller';
 import { MedicinesService } from './medicines.service';
@@ -9,10 +10,12 @@ import { DrugbankMedicinesService } from './sources/drugbank-medicines.service';
   controllers: [MedicinesController],
   providers: [
     MedicinesService,
+    MedicinesCacheAdminService,
     MedicinesCacheService,
     DrugbankMedicinesService,
     CnMedicinesService,
   ],
+  exports: [MedicinesCacheAdminService],
 })
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class -- NestJS @Module requires a class declaration
 export class MedicinesModule {}